Skeleton
Added some skeleton functions that need to be made probably. Also wrote an equation relating homography matrices so we find position of bot relative to world (should be double checked). Imported brains as a library and defined it into transmitToArduino (still needs to be fixed so that it will actually transmit an argument)